Add `Standard` prism element
This replaces `String`'s conformance to `PrismElement` with a dedicated wrapper in `Standard`. There were just too many issues involving with trying to make `String` conform, chief among them being that it's impossible to store values in extensions. The (very hacky and not at all stable) previous solution was to basically store everything in backing dictionaries, which led to values being retained and no longer being used. This puts that issue to bed.

In order to make the transition as seamless as possible, `ElementBuilder` has been expanded to accept instances of `String` in addition to conformers to `PrismElement`. This simply works as syntax sugar for wrapping the strings in `Standard` instances. For example, the following two prisms are semantically identical:

```swift
let prism1 = Prism {
    Standard("Hello, world!")
}

let prism2 = Prism {
    "Hello, world!"
}
```

This commit also fixes a bogus test that, for some reason, was (falsely) passing in the previous implementation, but (correctly) fails here.